+++ src/qiboconnection/job_listing.py
@@ -1,5 +1,4 @@
 """ JobListing class """
-# TODO: properly document these new code and the old one!
 
 from dataclasses import asdict, dataclass, field
 from datetime import datetime
@@ -13,7 +12,7 @@ from qiboconnection.util import decode_jsonified_dict
 
 @dataclass
 class JobListingItem:
-    """JobListing single item representation"""
+    """JobListing single item representation."""
 
     user_id: int
     device_id: int
@@ -21,8 +20,6 @@ class JobListingItem:
     job_type: str | JobType
     number_shots: int
     id: int | None = field(default=None)
-    # TODO: including created_at attribute requires modifying .execute() method
-    # created_at: datetime | None = field(default=None)
 
     @classmethod
     def from_response(cls, response: ListingJobResponse):